extracted mp4 previewer functionality from clip classifier.
